+++ src/AutoNumericOptions.js
@@ -28,6 +28,7 @@
  */
 
 import AutoNumeric from './AutoNumeric';
+import AutoNumericHelper from './AutoNumericHelper';
 
 /**
  * Options values enumeration
@@ -837,10 +838,14 @@ function freezeOptions(options) {
         if (optionName === 'valuesToStrings') {
             const vsProps = Object.getOwnPropertyNames(options.valuesToStrings);
             vsProps.forEach(valuesToStringObjectName => {
-                Object.freeze(options.valuesToStrings[valuesToStringObjectName]);
+                if (!AutoNumericHelper.isIE11() || options.valuesToStrings[valuesToStringObjectName] !== null) {
+                    Object.freeze(options.valuesToStrings[valuesToStringObjectName]);
+                }
             });
         } else if (optionName !== 'styleRules') {
-            Object.freeze(options[optionName]);
+            if (!AutoNumericHelper.isIE11() || options[optionName] !== null) {
+                Object.freeze(options[optionName]);
+            }
         }
     });